requirements.txt
pyyaml
faiss-cpu
sentence-transformers
toml